[#4093] Perform case-insensitive email search for existing users
When adding a new employment for a user, we look for the user based on the
email ID entered in the invitation dialog. This look-up was buggy in cases
where the email ID in the DB had non-lower case letters, since an exact
match search was being performed using lower cased emails entered in the
invitation dialog. This commit makes the search case-insensitive.

Fixes  #4093