[#4101] Generate the same thumbnail path irrespective of storage
`sorl` generated a different thumbnail path based on the storage. For our
use case of moving the cached images to Google Storage and re-using them,
this doesn't work very well, since we regenerate thumbnails for images that
already have thumbnails. This commit fixes that by using the same paths
generated using the filesystem storage backend, for the google one as well.

Closes #4101