[#3839] Add ability to re-run the survey to update data changes
Projects that have been created from a previous import of the survey can now
be updated with an updated survey CSV export by re-running the importer.

This allows us to avoid having to do "bulk" manual changes, as long as
survey field names are not the things that have been changed!