[#4146] Don't run signals when loading fixtures
`loaddata` calls the pre and post save hooks on the model. But, since
the related models may not be loaded, the signal handler may fail. The
docs [1] recommend this way of disabling the handlers. Ideally, we only
need the pre/post save handlers to have this code, but it's just easier
to have this everywhere (and may be even make it a re-usable decorator),
instead of forgetting to add it in some cases, etc.

[1]: https://docs.djangoproject.com/en/1.11/ref/django-admin/#loaddata