[#4148] Use pip-compile to manage requirements
This wouldn't update any of the requirements which have been locked
previously. This would make it easier to manage these requirements in
the future by listing only the direct requirements, and locking specific
libraries to versions which work for us, and letting other packages
upgrade to the latest version.

Specifically, we'll need to lock Django applications to versions that
are compatible with our version of Django. Other packages like Pillow,
lxml, etc. can easily be on their latest versions.