[#4196] Avoid repeated DB calls for project edit permissions
When we know we are calling the function User.can_edit_project on a list
of projects, it is much more efficient to compute the full list of
projects where the user has edit permissions, and then checking if the
current project is present in that list or not.

We could move to this model of permissions checking for checks too.
It is probably worth looking into our predicates to see if we can
include the optimisation at that level.