[#4202] Remove /request_organisation end-point
We no longer wish to support the workflow for requesting an employment
with an organisation. We remove this end-point, even though the mobile
app uses this end-point. The mobile app would show a 404 error message,
with not a very pretty looking error message, but not a lot of users are
going to use this workflow, in any case.