[#4222] Clear project directory cache on create too
Previously, the project directory cache was being cleared only on
updates and deletes, but we want to clear it on creates too since we are
talking about creation of related objects here, not just projects. For
example, a new sector or organisation/partnership could be created on a
project, and that should clear the cache.