[#4256] Use lookup.byIso for country names
We previously used lookup.byInternet but the values for the Internet
code for some countries is not the same as the ISO country code for
those countries, for instance Venezuela. Due to some caching issues,
previously the country codes were returning country names, and this
motivated the use of `byInternet` look up function, instead of `byIso`.