Push docker images to registry on rsr-env-* branches
To make it easier to deploy custom changes to training envs, we use a
temporary commit where we commit out some of the steps in the deploy
script. This change makes it easier to push the required docker images,
using branch names starting with `rsr-env-`.

This change makes the process less error prone and more convenient.