[#3810] Fix buggy is ME manager check for project
The bug was introduced in 5624da962ba41f7ba3be91d88aea91fd69209bde. The
call for `has_perm` was incorrect and didn't pass in the project as an
argument. So, any user who had M&E manager permissions on any project
was marked as M&E manager on every project they had access to. This
incorrectly gave the user M&E manager permissions on projects where they
are not supposed to have such permissions.

This bug was introduced in a commit that is quite old, but it didn't
manifest as a user visible problem until now, since it was not possible
to have different roles for users on projects, until the FAC feature was
built. Given that only Nuffic is using FAC currently, and they have
reported the problem, there shouldn't have been many instances of this
bug causing problems in the wild, yet.