[#4337] Load project directory at / always
Load project directory locally and in production, when we visit /.
Requires nginx changes for both these configurations. The redirect was
removed in favor of help from the browser, and a clean URL that feels
more shareable.

This commit also removes references to the old project directory URL
from the top bar and the org/update directories. The older redirect
views are not removed in order to not break browser history, etc. They
can be removed at some point in the future, though, once we are sure we
removed all the references to them internally, as well.