base.html: Ensure `request` is available in context
500.html inherits from base.html. When a REST view is broken and raises
a 500, this template is used. When this template is tried to be
rendered, request variable is not available in the rendering context and
this leads to an exception in the exception handler leading to recursive
errors that lead to a huge stack trace.

This commit fixes this problem by ensuring that the request variable is
available in context before trying to use it as an argument to a
template filter.