[#4345] Create IatiCheck objects using bulk create
Large projects can have 1000s of IatiCheck objects, especially since we
are creating objects for successes, warnings and errors. Creating these
objects in a loop can be terribly slow. And we call this method in the
project save, where we also have auto save. This means some projects can
put a lot of load on the DB.

The current commit switches to saving these objects using a bulk_create,
which significantly reduces the DB load for large projects.