[#4345] Filter queryset only for list view
We try to filter the queryset in PublicProjectViewSet, but this can be
slow. We only need to do this when the action is a list action, trying
to list all the projects (or related objects) that a user has
permissions for. When performing other operations like
read/update/delete, we don't need the permissions based filtering on the
queryset, since the individual object is checked for permissions.