[#4355] Don't fail iati_activity_id unique validation for ''
When an API request is made to /rest/v1/project/<pk> with an empty
string value for iati_activity_id, if there is another project in the DB
with an empty string set, we get a unique value error that is not
handled correctly by DRF validations.

This commit moves the code to set empty iati_activity_id fields to None
from Project.clean to Project.save so that it is run when updating a
project from the API as well.

Closes #4355