[#4356] Create/Update targets using dimension_value+period as "key"
We allow creating a disaggregation target by specifying the dimension
value id and the period id. For updates, the previous commit required
the ID of the target to be specified. This commit makes the API easier
to use by allowing update and create to send the same kind of data, and
doing the right thing depending on whether a target was already created
or not.