[#4187] Update existing organisation custom fields
Previously, we only ensured that the organisation custom fields existed.
But, each update of the data can update the dropdown options, and we
need to make sure that we update the organisation custom fields to
ensure that the drop down options are in sync with the project custom
field options. This wouldn't be required if we had a foriegn key
relationship between the project and the organisation custom fields.