[#4187] Explicitly specify the order of custom fields
Order the organisation custom fields by their order and pks when
returning them in the project directory API response. This would ensure
that the order of the filters in the dropdown created using these custom
fields can be modified by changing 'order' values in the DB, and it also
makes sure the order is deterministic.