[#4374] Add test to prevent manual changes to codelist JSONs
The codelist JSONs live in the project editor SPA, and could potentially
be edited manually by developers not aware how the files are generated.
This commit adds a test that prevents manual updates to these files,
informing developers how to generate the files properly.

We also skip coverage on akvo/codelists/scripts/ in this commit. The
code here was ignored in coverage before, since this was a scripts
directory. This commit converted the directory into a module, and hence
is no longer ignored by coverage without explicitly including it in the
omit list in .coveragerc.