[#4262] List potential enumerators in project enumerators end-point
We list potential enumerators for a project in
/project/<pk>/enumerators. This can be used to populate the enumerators
list for assignment. The enumerators listed here are users who have
explicitly been given an enumerator role on the project, or users
employed as Enumerators in one of the partner projects.

We don't worry about Enumerators in content-owned organisations, or
Enumerators on the root project of a project hierarchy for example. This
is a business rule that we enforce here to keep the logic for finding
enumerators simple.