[#4375] Add API end-point to send emails to enumerators
Each time an email is sent we update the token expiration time and also
increase the max_uses by 50. This should comfortably allow using the
token for a new project. The JSON data on the token stores information
about when the token was issued for which project. This can be used to
show information about when the last email was sent to a user.