[#4375] Explicitly use str(project_id) as key in RequestToken.data
We run into a weird bug where trying to update an object's JSON field
with a dict containing integer keys doesn't work correctly. The data
gets stored in the DB with a string key in the first instance. When the
json data is loaded as the object's field it has a string key, and then
trying to update that dict using dict.update, creates a new key with the
integer key. Now, when this is saved back into the DB, it looks like the
integer key is completely ignored, and the old value with the string key
is written back into the DB.

So, this commit explicitly switches to using str keys when trying to
write this data to the DB.