[#4559] Update store state when moving indicators/results
The previous code just updated the field array state and changed the UI to
reflect the move, but when a user interacted further with a moved item, the
change would appear to have been made on the old item at that index, because
the store state was not updated.

This change just changes the store state to reflect the truth, and does so
without changing how the API calls are made, etc. Ideally, we could possibly
improve this code further to make the API calls for the move from an action.