[#4624] Add Common Output Indicators report for EUTF
The report is similar to their existing program report with the full results
framework table, except that we filter for indicators related to a single
result.

In order to make the report somewhat configurable, we add the result ID to the
URL instead of hard-coding the result title to the report view.

The migration that creates the reports, though, uses the result name to add the
current result ID to the URL.