chore: Fix make-and-restore-production-dump.sh for clean envs
Fresh environments (linux at least) don't have all required directories or rights for said directories.
The default user had ID 1000 and the default gcloud image user has ID 100 --> no write rights
It's thus impossible for the user to write to the mounted volumes.

Additionally, access was given to **all** of `~/.config` which can be a security issue.
It's one of the default XDG configuration directories so many applications
 will store their config there and some might store unencrypted secrets.

Thus, this uses directories local to the project and keeps old backups, just in case.

#4650 Cannot dump and restore prod on fresh install