[#4656] Workaround performance of PublicProjectViewSet subclasses
The problematic method is _projects_filter_for_non_privileged_users which makes many calls to the DB (>2000).
We cache the response and try to invalidate it when users are added to removed from employment or a partnership.

Tests will have to be written and run to make sure things haven been broken.