Fix: Make react hot reload work again locally
I didn't want to mess around with JS code and point the hot reload in the client to another port.
It just isn't clear to me how to make that change a dev-only change.

The easiest route I therefore picked was to make it possible for the react client to access the ports
 it was expecting to access