feat: Add AkvoTreeModel.descendants(max_depth=...)
In order to test the abstract Model, stackoverflow provided a solution of creating a temporary table.
This uncovered a problem in test_results_framework.py where a child project should still have existed.

Subclasses of AkvoTreeModel now also don't have to set the path manually as a signal does it for all subclasses.

#4710: Improve project hierarchies