[#4947] IATI export datetime using correct format
The datetime stored in the DB was not aware of timezones and thus used the time on the server.

The last-updated-datetime attribute is an xsd:dateTime which can have a timezone component.
However, because the datetime instance is timezone naive and we use a manually formatted string,
 we exported the wrong time.

We now indicate to django that it should store timezone aware dates in the DB.
Additionally, old "last_modified_at" dates will be converted to timezone aware dates when exporting.