[#5017] Speed up Results framework endpoint
Prefetch as much data as possible.

It doesn't look like indicator.dimension_names is necessary so that was removed from the serializer.

`IndicatorPeriodFrameworkNoToLiteSerializer.get_updates` wasn't taking advantage of the prefetched data.
Now, it filters in python without making an extra query.
It's still suboptimal, but at least it's a bit faster.