Speed up indicator_data_period_framework endpoint
The UserDetailsSerializer adds many queries to get the approved_organisations.
From what I can see, the caller of POST/PATCH `indicator_data_period_framework` doesn't even consume the user data.
It would be great to remove it entirely, but it's not clear if the field is used elsewhere
 and the UserRawSerializer remove still provides some information.

After testing the Results admin to POST and PATCH, the operations took 1.5s instead of 16s.

Hopefully this doesn't break something else, elsewhere.