[#5022] Replace get_absolute_url with cacheable_url and cached_property in serializers and models
get_absolute_url uses `reverse()` which is filthy slow. Repeated calls to it slow down the application immensely.
We can cache that using `cached_property`, but optimally, we shouldn't be using such a method at all.

This is a quickfix to get use out of hot water. Hopefully, we can eradicate that method sometime...