[#5038] Fix enumerators can edit project updates they do not own
They aren't able to so anymore.

Anybody able to view the update was able to edit it.
That was changed by limiting the operation to the owners and
 introducing the "rsr.view_projectupdate" permission.

Because the permissions are now more limited, not everybody
 can change ProjectUpdates, so the expected number in
 test_permissions.py also had to change.

ProjectRole had to be taken into consideration too.
Normally, the "is_own" predicate would be ignored.
Unfortunately, there's no clean way to solve it, so
 an exception had to be made in the check for ProjectUpdates.