fix: internal IATI checks now run regardless of external one
Previously only a single check was run in order not surpass the rate limit of API calls to the external service.
This meant that the backlog was simply too long to run all internal tests in time.

We now reintroduce the use of Project.run_iati_checks in order to
 run the internal checks when the command `perform_iati_checks` is run in the cronjob.

#5096: Bug: IATI Export does not update checks after project is updated