[#5114] Fix import results set option not available in the editor
The import results set option will be shown/available if the current project has a parent.
In order to get the parent project, section5 only checks it by requesting endpoint related_project.
So, I added program property to identify whether it has a parent or not
when the response from that endpoint is empty and the project didn't set up its results framework yet.