[#5073] Fix Project editor - description section not loading
This bug is caused by all sections that has a dependency with the section1 not being re-fetching
when the ```editorRdr.section1.isFetched``` is true
due the conditional at the first time load is false.

So, I moved it in different area so that the all section will be executed once the```editorRdr`.section1.isFetched``` is ready.