refactor: Use root_period instead of program
The UI will have to easily find the jobs of child periods.
To make that easier, it's better to retrieve the root period from which one can get the root project.

It can also happen that there are master and sub programs.
The sub programs can also define their own periods that override those inherited by the parent program.
Searching for the parent period allows us to treat this case as well

#5145: Feature Request: UI for aggregation tasks