[#5199] Optimize /project/children endpoint
# Counting children

This was slow because we made a query for each project.
Now we get the children and grand children.
A simple dict with each project and a list of their children is then built in python.

**Side effect**

Paging doesn't work anymore as the queryset includes grand children.
Reimplementing it can be done another time if we really do have projects with enormous hierarchies.

# Parent retrieval

We already know who the parent is since we're getting its children --> no need to make another query.

# Locations

We weren't retrieving related objects, which we now do.
Additionally, we iterated over the list of countries twice when one iteration was enough.