[#5249] Add decorator to make django-q tasks unique
django-q doesn't support tracking of active task.
We therefore cannot make a simple query of "is this task already running".

All that is provided is a django signal before a task is executed with the task dictionary
 https://django-q.readthedocs.io/en/latest/signals.html#before-executing-a-task .
That's unfortunately not useful for our scenario as the edge case of an unexpected shutdown
 will block the unique task from executing in the future.
Were a timeout to be added and the task to still be running, the same task could be started again.

The chosen solution therefore, is a heartbeat thread.