[#5263] Improve docker-compose healthcheck conf
The `worker` service took a long time to start after the `web` service, because the interval for the `web` service was really long.
The interval was so high to cope with the long startup time of the `web` service, but the config option for that is `start_period` not `interval`.